+++ CMakeLists_files.cmake
@@ -141,7 +141,7 @@ list (APPEND PUBLIC_HEADER_FILES
 	opm/autodiff/fastSparseProduct.hpp
 	opm/autodiff/DuneMatrix.hpp
 	opm/autodiff/ExtractParallelGridInformationToISTL.hpp
-	opm/autodiff/flowMain.hpp
+	opm/autodiff/FlowMain.hpp
 	opm/autodiff/GeoProps.hpp
 	opm/autodiff/GridHelpers.hpp
 	opm/autodiff/GridInit.hpp
